backend/requirements.txt
# FastAPI 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Supabase 整合
supabase==2.3.0
python-dotenv==1.0.0

# LINE Bot SDK
line-bot-sdk==3.7.0

# 資料處理
python-dateutil==2.8.2

# 工具函式庫
requests==2.31.0

# 開發工具
pytest==7.4.3
black==23.11.0
flake8==6.1.0
